requirements.txt
pydantic
fastapi
uvicorn[standard]
fastapi[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0